+++ src/main/java/com/itcube/journal/repos/AttendanceRepo.java
@@ -1,11 +1,13 @@
 package com.itcube.journal.repos;
 
 import com.itcube.journal.domain.Attendance;
+import com.itcube.journal.domain.ThematicPlanning;
+import com.itcube.journal.domain.User;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface AttendanceRepo extends CrudRepository<Attendance, Long> {
+public interface AttendanceRepo extends CrudRepository<Attendance, Integer> {
 
     List<Attendance> findByStudents_Surname(String surname);
